refactor(eve): fold approval coordination into interpreter
